requirements.txt
# Core PyTorch & Hardware Optimization
torch>=2.1.0
torchvision
accelerate>=0.26.0
bitsandbytes>=0.41.0

# Vision-Language Models & Segmentation
transformers>=4.40.0
qwen-vl-utils
Pillow>=9.0.0

# Topological Math & Computer Vision
numpy>=1.24.0
opencv-python>=4.8.0

# NLP (The Latent Bridge)
spacy>=3.7.0

# Metrics & Artifact Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0